// block-performance-1.0.0/src/library/package-json/dependency-ls/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {}

export interface OutputOk {
    list: { package_name: string; version: string; type: "dependency" | "dev-dependency" }[];
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            throw new Error("v contains unknown field: " + field);
        });
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            list: copy_list(v.list)
        };
        return obj;
    } else {
        return { list: [] };
    }

    function copy_list(v: any): { package_name: string; version: string; type: "dependency" | "dev-dependency" }[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): { package_name: string; version: string; type: "dependency" | "dev-dependency" } {
            if (typeof v === "object" && v !== null) {
                const obj = {
                    package_name: copy_package_name(v.package_name),
                    version: copy_version(v.version),
                    type: copy_type(v.type)
                };
                return obj;
            } else {
                return { package_name: "", version: "", type: "dependency" };
            }

            function copy_package_name(v: any): string {
                return typeof v === "string" ? v : "";
            }

            function copy_version(v: any): string {
                return typeof v === "string" ? v : "";
            }

            function copy_type(v: any): "dependency" | "dev-dependency" {
                if (typeof v === "string" && new Set(["dependency", "dev-dependency"]).has(v) === true) {
                    return v as any;
                } else {
                    return "dependency" as any;
                }
            }
        }
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;